ref(new recipe) -  split code and use constant type for selection
